requirements.txt
requests
pyTelegramBotAPI
python-dotenv
pillow
openai
